SF-92
Went a little farther down the property path tree than I meant to, but it is a small amount of code, and easily re-arranged later.

This parsing line is not hooked up to a Query Engine piece yet.  I will work on writing the Unit Tests once we get there, because I am not exactly sure at what point in this chain I am supposed to be returning...  Also pathElt is just a stubbed out AMD module waiting for some code.