SF-177
Sparql Percent.  was able to reuse in turtle engine, but was unable to lazy load HEX parser at this time.  too many places where percent is appended to a string which is then passed back to another parser.  Perhaps we can make this work in the future, but not yet