SF-52
Created USING clause parser.  Returns an instance of the new Graph Loader object that should be used by USING and FROM statements.  Graph loader handles going back to a server to get non-NAMED graph data.  And by handles I mean it passes the work off to a fetch method that should be attached to the query options.