SF-107
Created Blank Node Property List parser..  Returns the bNode.

Refactor RdfJs Graph to take functions instead of object in it's array-like methods.
Reason is that in JavaScript forEach is supposed to take a function, and having to remember that it is an object in this one case is more error prone.

Might add a task later to make Graph even more Array-like by adding map, and any other useful Array prototype methods that look cool